Added index. Added main name choice.
Added the option to add an index.html file.
Added the option to specify the name of the Main file.
Made Main file name, module name and the initialization in index.html dynamic (depending on the choice of main file name)
Fixed default dependencies to include at least the core library.